requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
Pillow>=10.0.0

# AI Integration
openai>=1.10.0

# Screen Capture & Automation
mss>=9.0.0
pyautogui>=0.9.54
pydirectinput>=1.0.4  # Windows-specific, better for games

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0

# Packaging
pyinstaller>=6.3.0
